Correction and test implementation
