+++ ross/utils.py
@@ -11,7 +11,6 @@ from copy import deepcopy as copy
 from numba import njit
 from numpy.fft import fft
 import control as ct
-from rotor_assembly import Rotor
 
 
 class NumpyEncoder(json.JSONEncoder):
@@ -1497,57 +1496,3 @@ def is_scalar(parameter, parameter_name):
         raise ValueError(f"{parameter_name} must be a scalar.")
 
     return np.array(parameter)
-
-def concatenate_rotor(rotor_list):
-    shaft_elements = []
-    disk_elements = []
-    bearing_elements = []
-    point_mass_elements = []
-
-    node_offset = 0
-    rotor_id = 0  # Identificador incremental para tags
-
-    for rotor in rotor_list:
-        rotor = copy(rotor)
-
-        # Reindexar elementos de eixo
-        for i, el in enumerate(rotor.shaft_elements):
-            el.n_l += node_offset
-            el.n_r += node_offset
-            el.n = el.n_l  # importante para elementos do ROSS
-            el.tag = f"shaft_r{rotor_id}_{i}"
-        shaft_elements.extend(rotor.shaft_elements)
-
-        # Reindexar discos
-        for i, el in enumerate(rotor.disk_elements):
-            el.n += node_offset
-            el.tag = f"disk_r{rotor_id}_{i}"
-        disk_elements.extend(rotor.disk_elements)
-
-        # Reindexar mancais
-        for i, el in enumerate(rotor.bearing_elements):
-            el.n += node_offset
-            el.tag = f"bearing_r{rotor_id}_{i}"
-        bearing_elements.extend(rotor.bearing_elements)
-
-        # Reindexar massas concentradas
-        for i, el in enumerate(rotor.point_mass_elements):
-            el.n += node_offset
-            el.tag = f"pmass_r{rotor_id}_{i}"
-        point_mass_elements.extend(rotor.point_mass_elements)
-
-        # Atualiza deslocamento para o próximo rotor
-        all_nodes = [el.n_r for el in rotor.shaft_elements] + [
-            el.n for el in rotor.disk_elements + rotor.bearing_elements
-        ]
-        node_offset = max(all_nodes)
-        rotor_id += 1
-
-    rotor_concat = Rotor(
-        shaft_elements=shaft_elements,
-        disk_elements=disk_elements,
-        bearing_elements=bearing_elements,
-        point_mass_elements=point_mass_elements,
-    )
-
-    return rotor_concat
